move to plugin and add analytics sample rate
