+++ ddtrace/contrib/pytest/constants.py
@@ -1,0 +1,3 @@
+FRAMEWORK = "pytest"
+
+HELP_MSG = "Enable tracing of pytest functions."
